# Conflicts:
#	save/achievement.sqlite3

+++ core/src/main/java/com/apple/pawn/GameScreen.java
@@ -23,7 +23,6 @@ import com.badlogic.gdx.utils.ScreenUtils;
 import com.badlogic.gdx.utils.viewport.FitViewport;
 
 import java.util.Iterator;
-import java.util.Objects;
 import java.util.function.IntSupplier;
 
 /**
@@ -63,6 +62,7 @@ public class GameScreen implements Screen {
 	private int turnCount;
 	private String order;
 	private final Sprite cursor;
+	private boolean isLoad;					// 続きからのゲームか
 
 	//---- 他のクラス
 	private GameSetting gameSetting;				// ゲームの設定
@@ -96,6 +96,7 @@ public class GameScreen implements Screen {
 		mapCameraHeight = (float) BoardSurface.MAP_HEIGHT / 2;
 		mapCameraWidth = (float) BoardSurface.MAP_WIDTH / 2;
 		turnCount = 1;
+		isLoad = false;
 
 		//---- カメラ関係の初期化
 		camera = new OrthographicCamera();
@@ -181,6 +182,7 @@ public class GameScreen implements Screen {
 	}
 
 	public void load(final SaveData sd) {
+		isLoad = true;
 		gameSetting = new GameSetting();
 		gameSetting.setStageNo(sd.mapNo);
 		board.initialize(manager,sd.mapNo,font);
@@ -212,14 +214,14 @@ public class GameScreen implements Screen {
 		manager.update();
 
 		//------ 入力
-		if (Gdx.input.isKeyPressed(Input.Keys.F1)) {
+		if (Gdx.input.isKeyJustPressed(Input.Keys.F1)) {
 			game.setScreen(new TitleScreen(game));
 		}
-		if (Gdx.input.isKeyPressed(Input.Keys.F5)) {
+		if (Gdx.input.isKeyJustPressed(Input.Keys.F5)) {
 			saveData.setGameState(timer, gameSetting.getStageNo(), goalNo, sequenceNo, turnPlayerNo, turnCount);
 			fileIO.save();
 		}
-		if (Gdx.input.isKeyPressed(Input.Keys.F6)) {
+		if (Gdx.input.isKeyJustPressed(Input.Keys.F6)) {
 			fileIO.load();
 			GameScreen gameScreen = new GameScreen(game);
 			gameScreen.load(fileIO.getSaveData());
@@ -539,7 +541,7 @@ public class GameScreen implements Screen {
 			if(timer-timerRap >= 0.5f) sequenceNo+=2;
 		}
 		if(sequenceNo == Sequence.TASK_DO.no +4) {
-			if(timer-timerRap >= 5.0f) sequenceNo++;
+			if(timer-timerRap >= 3.0f) sequenceNo++;
 		}
 		if(sequenceNo == Sequence.TASK_DO.no +5) {
 			// ゲーム情報の更新
@@ -553,7 +555,9 @@ public class GameScreen implements Screen {
 
 	private int result() {
 		if(sequenceNo == Sequence.RESULT.no) {
-			fileIO.delete();
+			if(isLoad) {
+				fileIO.delete();
+			}
 			StringBuilder txt = new StringBuilder("全員ゴールしたよ");
 			Iterator<Player> playerIterator = new Array.ArrayIterator<>(playerManager.getGoalPlayer());
 			while(playerIterator.hasNext()) {
